% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspace2-package.R
\docType{package}
\name{aspace2-package}
\alias{aspace2}
\alias{aspace2-package}
\title{aspace2: Centrographic statistics and computational geometries for spatial point patterns in 2D/3D}
\description{
aspace2 package is a fork of aspace package from Randy Bui, Ron N. Buliung and Tarmo K. Remmel (2012). Like these, aspace2 package contains the same collection of functions for computing centrographic statistics (e.g., standard distance, standard deviation ellipse, standard deviation box) for observations taken at point locations in 2D or 3D. The aspace2 library was conceived to avoid the default output assigns and file writings. This is primarilly intended to be used in a tidy* context, (e.g in `dplyr::groupy_by`, `dplyr::summarise`, `dplyr::mutate`, etc, contexts).
}
\author{
\strong{Maintainer}: Gabriel V. Gaona \email{gabo@gavg712.com} (\href{https://orcid.org/0000-0003-4666-2248}{ORCID})

}
\keyword{internal}
