% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{median_centre}
\alias{median_centre}
\alias{mean_centre}
\alias{central_feature}
\alias{centre_min_dis}
\alias{central_feature2pts}
\title{Centrality calculator in 2D or 3D}
\usage{
median_centre(points)

mean_centre(points, weights = NULL)

central_feature(points)

centre_min_dis(points, dist = 100)

central_feature2pts(points, points2)
}
\arguments{
\item{points, points2}{matrix, datat.frame or tibble of points 2D or 3D}

\item{weights}{Numeric. Same length of number of points.}

\item{dist}{Atomic numeric, Default 100. Hold distance value between i
and ith iterations.}
}
\value{
Depends on input, returns a Numeric vector of 2 or 3 values
}
\description{
Centrality calculator in 2D or 3D
}
\details{
Central minimum distance can only performed in 2D. Other centrality
  calculations could be used in 2D or 3D points
}
\note{
Mean, Median and Central features calculators are modified versions
  of `aspace::*()` from Ron Buliung & Randy Bui (2012)
}
\author{
Gabriel Gaona
}
