% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{distances}
\title{Euclidean distance calculator in 2D or 3D}
\usage{
distances(destmat, centre = NULL)
}
\arguments{
\item{destmat}{matrix, datat.frame or tibble of target points 2D or 3D}

\item{centre}{Numeric. Coordinates 2D or 3D of central point.}
}
\value{
Depends on input, returns a Numeric vector of distances to each
  location from the specified centre.
}
\description{
Euclidean distance calculator in 2D or 3D
}
\details{
Given a matriz of xy or xyz coordinates for a set of point locations
  compute the euclidean distances from a centre coordinate. This can be used
  in a planar coordinate system (e.g. projected coordinates in UTM)
}
\author{
Gabriel Gaona
}
