% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_box.R
\name{sd_box}
\alias{sd_box}
\title{Standard deviation box calculator in 2D or 3D}
\usage{
sd_box(points, centre = NULL, weights = NULL, output = c("coord", "param"))
}
\arguments{
\item{points}{matrix, datat.frame or tibble of points 2D or 3D}

\item{centre}{Numeric. Coordinates 2D or 3D of central point. Default NULL,
performs a calculation of mean_centre() from point localities}

\item{weights}{Numeric. Same length of number of points.}

\item{output}{Character. "coord" to return coordinates or sd box in 2D or 3D.
"param" to return sd box attributes 2D or 3D. Default c("coord", "param"),
to return both as list.}
}
\value{
Depends on input, "coords" returns a data.frame of 2 or 3 columns and
  4 or 8 point coordinates. "param" returns a data.frame with centre
  coordinates, standard deviation in each axis, space(area for 2D, volume for
  3D) and number of dimensions in coordinates.
}
\description{
Standard deviation box calculator in 2D or 3D
}
\author{
Gabriel Gaona
}
