% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_distance.R
\name{st_sd_distance}
\alias{st_sd_distance}
\alias{st_sd_distance.sfg}
\alias{st_sd_distance.sf}
\alias{st_sd_distance.sfc}
\title{Standard deviation distance calculator (2D)}
\usage{
st_sd_distance(.x, centre = NULL, weights = NULL, ...)

\method{st_sd_distance}{sfg}(.x, centre = NULL, weights = NULL, ...)

\method{st_sd_distance}{sf}(.x, centre = NULL, weights = NULL, ...)

\method{st_sd_distance}{sfc}(.x, centre = NULL, weights = NULL, ...)
}
\arguments{
\item{.x}{Points \code{sf} object}

\item{centre}{\code{sf} \code{sfc} of one point, numeric, matrix, data.frame
or list of central point. Default NULL, means a calculation of the "mean_centre"
from \code{.x} localities.}

\item{weights}{Numeric. Same length as number of points.}

\item{...}{other parameters for \code{sf::st_distance()}}
}
\value{
A sf \code{"POLYGON"} with atributes: radius (standard deviation distance)
and area surrounding, perimeter, center coordinates, and if weigted indicator.
}
\description{
Standard deviation distance calculator (2D)
}
\examples{
  requireNamespace("ggplot2", quietly = TRUE)
  library(sf, quietly = TRUE)
  library(ggplot2)
  bbx <- matrix(c(697047,9553483,
                  696158,9560476,
                  700964,9561425,
                  701745,9555358),
                byrow = TRUE,
                ncol = 2)
  bbx <- st_multipoint(bbx)
  bbx <- st_cast(bbx,"POLYGON")
  bbx <- st_sfc(bbx, crs = 31992)
  set.seed(1234)
  points <- st_sf(geometry = st_sample(bbx, 100))
  SDD <- st_sd_distance(points)
  ggplot() +
    geom_sf(data = SDD, fill = NA, color = "darkolivegreen") +
    geom_sf(data = points, color = "steelblue", size = 0.5)
}
\author{
Gabriel Gaona
}
