% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{sin_d}
\alias{sin_d}
\alias{cos_d}
\alias{tan_d}
\alias{asin_d}
\alias{acos_d}
\alias{atan_d}
\alias{as_radians}
\title{Trigonometric functions to compute in degrees}
\usage{
sin_d(x)

cos_d(x)

tan_d(x)

asin_d(x)

acos_d(x)

atan_d(x)

as_radians(x)
}
\arguments{
\item{x}{Numeric vector of angles in degrees}
}
\value{
Numeric vector of trigonometric computation
}
\description{
To simplify the use of radians in R
}
\author{
Gabriel Gaona
}
\keyword{as_radians}
\keyword{degrees,}
\keyword{trigonometry,}
