% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_distance.R
\name{sd_distance}
\alias{sd_distance}
\title{Standard deviation distance calculator (2D)}
\usage{
sd_distance(
  points = NULL,
  x = NULL,
  y = NULL,
  centre = NULL,
  weights = NULL,
  output = c("coord", "param")
)
}
\arguments{
\item{points}{matrix, datat.frame or tibble of points 2D}

\item{x, y}{Optional. numeric vectors for `x` and `y` coordinates of points.
Default NULL perform extraction from point localities (`x = points[,1]` and
`y <- points[,2]`)}

\item{centre}{Numeric. Coordinates 2D of central point. Default NULL,
performs a calculation of `mean_centre()` from point localities}

\item{weights}{Numeric. Same length of number of points.}

\item{output}{Character. `"coord"` to return coordinates for sd_distance
points in 2D. "param" to return sd_distance attributes for 2D. Default
c("coord", "param") to return both as list.}
}
\value{
Depends on input, "coords" returns a data.frame of 2 and 360 point
  coordinates. "param" returns a data.frame with weigted indicator (`1 = TRUE`,
  or `0 = FALSE`), centre coordinates, standard deviation distance radius, and
  area of sd_distance.
}
\description{
Standard deviation distance calculator (2D)
}
\author{
Gabriel Gaona
}
